[Profiling] Fix compile error on MacOS (#26854)
This fixes the following template instantiation error on MacOS for me,
which occurs only with `cmake-modular-profiling`:

```text
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:265:9: error: no matching constructor for initialization of 'M::ProfilingDetail::Label'
        detail(std::forward<DetailStr>(detail), detailValue) {}
        ^      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/Users/bduke/work/modular/third-party/llvm-project/llvm/include/llvm/ADT/SmallVector.h:945:33: note: in instantiation of function template specialization 'M::ProfilingDetail::BeginEvent::BeginEvent<llvm::StringLiteral, unsigned long &>' requested here
    ::new ((void *)this->end()) T(std::forward<ArgTypes>(Args)...);
                                ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:344:25: note: in instantiation of function template specialization 'llvm::SmallVectorImpl<M::ProfilingDetail::BeginEvent>::emplace_back<unsigned long long, unsigned long &, unsigned long &, llvm::StringLiteral, unsigned long &>' requested here
    return last->events.emplace_back(std::forward<Args>(args)...);
                        ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:453:60: note: in instantiation of function template specialization 'M::ProfilingDetail::EventList<M::ProfilingDetail::BeginEvent>::emplace_back<unsigned long long, unsigned long &, unsigned long &, llvm::StringLiteral, unsigned long &>' requested here
    [[maybe_unused]] const BeginEvent &event = beginEvents.emplace_back(
                                                           ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:734:16: note: in instantiation of function template specialization 'M::ProfilingDetail::TimeTraceThreadProfiler::beginWithParent<llvm::StringLiteral, unsigned long &>' requested here
          ctx->beginWithParent(parentId, std::forward<Args>(args)...));
               ^
/Users/bduke/work/modular/KGEN/lib/CompilerRT/OutputChain.cpp:152:49: note: in instantiation of function template specialization 'M::ProfilerEntry<true>::createWithParent<llvm::StringLiteral, unsigned long &>' requested here
        TimeTraceScope scope(MojoProfilerEntry::createWithParent(
                                                ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:150:13: note: candidate constructor not viable: no known conversion from 'unsigned long' to 'llvm::StringLiteral' for 1st argument
  constexpr Label(StringLiteral stringLiteral, uint64_t value = kNoIntPayload)
            ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:157:13: note: candidate constructor not viable: no known conversion from 'unsigned long' to 'M::InternableString' for 1st argument
  constexpr Label(InternableString internableString,
            ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:165:3: note: candidate constructor not viable: no known conversion from 'unsigned long' to 'std::string' (aka 'basic_string<char, char_traits<char>, allocator<char>>') for 1st argument
  Label(std::string string, uint64_t value = kNoIntPayload)
  ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:172:3: note: candidate constructor not viable: no known conversion from 'unsigned long' to 'llvm::StringRef' for 1st argument
  Label(StringRef stringRef, uint64_t value = kNoIntPayload)
  ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:179:3: note: candidate constructor not viable: no known conversion from 'unsigned long' to 'M::ProfilerPrintFn' (aka 'function_ref<basic_string<char, char_traits<char>, allocator<char>> ()>') for 1st argument
  Label(ProfilerPrintFn printFn, uint64_t value = kNoIntPayload)
  ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:143:13: note: candidate constructor not viable: allows at most single argument 'value', but 2 arguments were provided
  constexpr Label(uint64_t value = kNoIntPayload)
            ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:199:3: note: candidate constructor not viable: requires 1 argument, but 2 were provided
  Label(const Label &) = delete;
  ^
/Users/bduke/work/modular/Support/include/Support/Profiling/TimeProfiler.h:201:3: note: candidate constructor not viable: requires single argument 'that', but 2 arguments were provided
  Label(Label &&that) { moveFrom(that); }
```

MODULAR_ORIG_COMMIT_REV_ID: 95179e2d9e835280326afb689f818083570e53dd

+++ KGEN/lib/CompilerRT/OutputChain.cpp
@@ -150,7 +150,7 @@ void OutputChain::executeAsTask(void (*resume)(int8_t *), int8_t *hdl,
       [parentId = this->parentEventId, taskId, resume, hdl]() mutable {
         // Use the 'prototype' profiling entry, but augment with the task id.
         TimeTraceScope scope(MojoProfilerEntry::createWithParent(
-            parentId, StringLiteral("task"), taskId));
+            parentId, StringLiteral("task"), (uint64_t)taskId));
         resume(hdl);
 #if MODULAR_PARANOID
         // Sleeping here gives any await loop the chance to exit and
